+++ packages/fuselage/src/components/Box/Scrollable/index.js
@@ -1,21 +1,25 @@
-import { useClassName } from '@rocket.chat/fuselage-hooks';
 import PropTypes from 'prop-types';
 import React from 'react';
 
 import { useProps } from '../../../hooks';
 
-export function Scrollable({ children, horizontal, vertical }) {
+export function Scrollable({ children, horizontal, vertical, smooth }) {
   const [, PropsProvider] = useProps(({ className, ...props }) => ({
-    className: useClassName('rcx-scrollable', { horizontal, vertical }, className),
+    className: [
+      className,
+      'rcx-box--scrollable',
+      horizontal && 'rcx-box--scrollable-horizontal',
+      vertical && 'rcx-box--scrollable-vertical',
+      smooth && 'rcx-box--scrollable-smooth',
+    ].filter(Boolean).join(' '),
     ...props,
-  }));
+  }), [horizontal, vertical, smooth]);
 
   return <PropsProvider children={children} />;
 }
 
-Scrollable.displayName = 'Scrollable';
-
 Scrollable.propTypes = {
   horizontal: PropTypes.bool,
   vertical: PropTypes.bool,
+  smooth: PropTypes.bool,
 };