+++ packages/fuselage/src/components/Chip/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Avatar } from '../Avatar';
-import { Box, Flex, Margins, MarginsWrapper } from '../Box';
+import { Box, Flex, Margins } from '../Box';
 import { Icon } from '../Icon';
 
 const Wrapper = Box.extend('rcx-chip__wrapper', 'div');
@@ -20,16 +20,16 @@ export const Chip = ({
   ...props
 }) => (
   <Flex.Container>
-    <Container disabled={!(onClick || onMouseDown)} onClick={onClick || onMouseDown} {...props}>
-      {Thumb && thumbUrl && <Margins all={4}>
+    <Container type='button' disabled={!(onClick || onMouseDown)} onClick={onClick || onMouseDown} {...props}>
+      {Thumb && thumbUrl && <Margins all='x4'>
         <Thumb url={thumbUrl} />
       </Margins>}
       {children && <Flex.Item shrink={1}>
-        <Margins all={4}>
+        <Margins all='x4'>
           <Box is='span' textStyle='p1' textColor='default' className='rcx-chip__text'>{children}</Box>
         </Margins>
       </Flex.Item>}
-      {Remove && (onClick || onMouseDown) && <Margins all={4}>
+      {Remove && (onClick || onMouseDown) && <Margins all='x4'>
         <Box>
           <Remove/>
         </Box>
@@ -41,16 +41,16 @@ export const Chip = ({
 Chip.displayName = 'Chip';
 
 Chip.Wrapper = ({ children, width, alignItems = 'center', wrap = 'wrap', ...props }) =>
-  <MarginsWrapper all={4}>
+  <Margins all='neg-x4'>
     <Flex.Container alignItems={alignItems} wrap={wrap}>
       <Wrapper {...props}>
         {children.map((children, i) =>
           <Flex.Item key={i} shrink={1}>
-            <Margins all={4}>
+            <Margins all='x4'>
               {children}
             </Margins>
           </Flex.Item>,
         )}
       </Wrapper>
     </Flex.Container>
-  </MarginsWrapper>;
+  </Margins>;