+++ packages/fuselage/src/components/Grid/index.js
@@ -3,18 +3,21 @@ import React from 'react';
 
 import { Box } from '../Box';
 
+const GridWrapper = Box.extend('rcx-grid__wrapper');
 const StyledGrid = Box.extend('rcx-grid');
 const StyledGridItem = Box.extend('rcx-grid__item');
 
 export function Grid({ xs, sm, md, lg, xl, ...props }) {
-  return <StyledGrid
-    mod-xs={xs}
-    mod-sm={sm}
-    mod-md={md}
-    mod-lg={lg}
-    mod-xl={xl}
-    {...props}
-  />;
+  return <GridWrapper>
+    <StyledGrid
+      mod-xs={xs}
+      mod-sm={sm}
+      mod-md={md}
+      mod-lg={lg}
+      mod-xl={xl}
+      {...props}
+    />
+  </GridWrapper>;
 }
 
 Grid.displayName = 'Grid.Row';