+++ packages/fuselage/src/components/index.js
@@ -37,3 +37,5 @@ export * from './AutoComplete';
 export * from './Options';
 export * from './Select';
 export * from './Tooltip';
+export * from './Modal';
+export * from './Loading';