+++ packages/ui-kit/src/index.ts
@@ -17,6 +17,7 @@ export enum ELEMENT_TYPES {
   CONTEXT= 'context',
   FIELDS= 'fields',
   INPUT= 'input',
+  PLAIN_TEXT='plain_text',
   TEXT='text',
   MARKDOWN='mrkdwn'
 };
@@ -42,30 +43,40 @@ export interface UiKitText extends UiKitElement {
   text: string;
 }
 
-export abstract class UiKitParserMessage {
-  text: (text: UiKitText, context) => Component;
-  button: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
-  image: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
-  datePicker: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
-  staticSelect: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
-  multiStaticSelect: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
-  selectInput: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
-  context: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
-  divider: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
-  actions: (element: UiKitElement, context: BLOCK_CONTEXT) => Component;
+export abstract class UiKitParserButtons {
+  button: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+}
+
+export abstract class UiKitParserText {
+  text: (text: UiKitText, context: BLOCK_CONTEXT, index: Number) => Component;
+  plaintText: (text: UiKitText, context: BLOCK_CONTEXT, index: Number) => Component;
+  mrkdwn: (text: UiKitText, context: BLOCK_CONTEXT, index: Number) => Component;
+}
+
+export abstract class UiKitParserMessage extends UiKitParserText {
+  button: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  image: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  datePicker: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  staticSelect: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  multiStaticSelect: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  // selectInput: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  context: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  divider: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  actions: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
+  overflow: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component;
 
   renderAccessories = createRenderElement([
-    ELEMENT_TYPES.IMAGE,
     ELEMENT_TYPES.BUTTON,
-    ELEMENT_TYPES.STATIC_SELECT,
+    ELEMENT_TYPES.IMAGE,
     ELEMENT_TYPES.MULTI_STATIC_SELECT,
+    ELEMENT_TYPES.STATIC_SELECT,
     ELEMENT_TYPES.CONVERSATION_SELECT,
+    ELEMENT_TYPES.USER_SELECT,
     ELEMENT_TYPES.CHANNEL_SELECT,
     ELEMENT_TYPES.USER_SELECT,
-    ELEMENT_TYPES.USER_SELECT,
     ELEMENT_TYPES.DATEPICKER,
+    ELEMENT_TYPES.OVERFLOW,
   ]);
-
 
   renderActions = createRenderElement([
     ELEMENT_TYPES.BUTTON,
@@ -86,7 +97,7 @@ export abstract class UiKitParserMessage {
 }
 
 export abstract class UiKitParserModal extends UiKitParserMessage {
-  plainInput: (element: UiKitElement, context: BLOCK_CONTEXT) => Component
+  plainInput: (element: UiKitElement, context: BLOCK_CONTEXT, index: Number) => Component
   renderInputs = createRenderElement([
     ELEMENT_TYPES.STATIC_SELECT,
     ELEMENT_TYPES.PLAIN_TEXT_INPUT,
@@ -99,46 +110,58 @@ export abstract class UiKitParserModal extends UiKitParserMessage {
   ]);
 }
 
-export const uiKitGeneric = <T> (allowedItems?: Array<ELEMENT_TYPES>)  => (parser : T) => ({ blocks = [] }) =>
+export const uiKitGeneric = <T> (allowedItems?: Array<ELEMENT_TYPES>)  => (parser : T) => (blocks) =>
   blocks
     .filter(({ type }) => !allowedItems || allowedItems.includes(type))
-    .map(({ type, ...block }: UiKitElement) => parser[type] ? parser[type](block, BLOCK_CONTEXT.BLOCK, parser) : type );
+    .map(({ type, ...block }: UiKitElement, i) => parser[type] ? parser[type](block, BLOCK_CONTEXT.BLOCK, i) : type );
 
 
-const renderElement = ({ type, ...element }: UiKitElement, context: BLOCK_CONTEXT, parser: UiKitParser) =>  {
+const renderElement = ({ type, ...element }: UiKitElement, context: BLOCK_CONTEXT, parser: UiKitParser, index: Number) =>  {
   switch (type as ELEMENT_TYPES) {
+    case ELEMENT_TYPES.OVERFLOW:
+      return parser.overflow({ type, ...element } as UiKitText, context, index);
     case ELEMENT_TYPES.MARKDOWN:
     case ELEMENT_TYPES.TEXT:
-      return parser.text({ type, ...element } as UiKitText, context);
+      return parser.text({ type, ...element } as UiKitText, context, index);
     case ELEMENT_TYPES.BUTTON:
-      return parser.button(element as UiKitElement, context);
+      return parser.button(element as UiKitElement, context, index);
     case ELEMENT_TYPES.IMAGE:
-      return parser.image(element as UiKitElement, context);
+      return parser.image(element as UiKitElement, context, index);
     case ELEMENT_TYPES.STATIC_SELECT:
-      return parser.staticSelect(element as UiKitElement, context);
+      return parser.staticSelect(element as UiKitElement, context, index);
     case ELEMENT_TYPES.MULTI_STATIC_SELECT:
-      return parser.multiStaticSelect(element as UiKitElement, context);
+      return parser.multiStaticSelect(element as UiKitElement, context, index);
     case ELEMENT_TYPES.DATEPICKER:
-      return parser.datePicker(element as UiKitElement, context);
+      return parser.datePicker(element as UiKitElement, context, index);
     case ELEMENT_TYPES.PLAIN_TEXT_INPUT:
-      return parser.plainInput(element as UiKitElement, context);
-    case ELEMENT_TYPES.CONVERSATION_SELECT:
-    case ELEMENT_TYPES.CHANNEL_SELECT:
-    case ELEMENT_TYPES.USER_SELECT:
-      return parser.selectInput({ type, ...element }, context);
+      return parser.plainInput(element as UiKitElement, context, index);
+    // case ELEMENT_TYPES.CONVERSATION_SELECT:
+    // case ELEMENT_TYPES.CHANNEL_SELECT:
+    // case ELEMENT_TYPES.USER_SELECT:
+    //   return parser.selectInput({ type, ...element }, context, index);
   }
 };
 
-export const createRenderElement = (allowedItems?: Array<ELEMENT_TYPES>) => (element: UiKitElement, context: BLOCK_CONTEXT, parser) => {
+export const createRenderElement = (allowedItems?: Array<ELEMENT_TYPES>) => (element: UiKitElement, context: BLOCK_CONTEXT, parser, index) => {
   if (allowedItems && !allowedItems.includes(element.type)) {
     return null;
   }
-  return renderElement(element, context, parser);
+  return renderElement(element, context, parser, index);
 };
 
+export const uiKitButtons = uiKitGeneric<UiKitParserButtons>([
+  ELEMENT_TYPES.BUTTON,
+]);
+
+export const uiKitText = uiKitGeneric<UiKitParserText>([
+  ELEMENT_TYPES.TEXT,
+  ELEMENT_TYPES.PLAIN_TEXT,
+  ELEMENT_TYPES.MARKDOWN,
+]);
+
 export const uiKitMessage = uiKitGeneric<UiKitParserMessage>([
-  ELEMENT_TYPES.SECTION,
   ELEMENT_TYPES.DIVIDER,
+  ELEMENT_TYPES.SECTION,
   ELEMENT_TYPES.IMAGE,
   ELEMENT_TYPES.ACTIONS,
   ELEMENT_TYPES.CONTEXT