+++ packages/fuselage/CHANGELOG.md
@@ -3,6 +3,121 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.30](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.29...v0.2.0-alpha.30) (2020-02-12)
+
+
+### Bug Fixes
+
+* Safari not triggering buttons anchors ([#146](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/146)) ([2cb5aaa](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/2cb5aaaf4c71e277a0a8ea556a8b9efbb3e58620))
+
+
+
+
+
+# [0.2.0-alpha.29](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.28...v0.2.0-alpha.29) (2020-02-10)
+
+**Note:** Version bump only for package @rocket.chat/fuselage
+
+
+
+
+
+# [0.2.0-alpha.28](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.27...v0.2.0-alpha.28) (2020-02-10)
+
+
+### Bug Fixes
+
+* placeholder transition ([#144](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/144)) ([4ce91d1](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/4ce91d179ece15d9014b93eb878b644e79850f56))
+
+
+
+
+
+# [0.2.0-alpha.26](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.25...v0.2.0-alpha.26) (2020-02-10)
+
+
+### Bug Fixes
+
+* Select autofocus ([#141](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/141)) ([25849ee](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/25849eed55e4edbf26f54ae4a72c71c1d528e850))
+
+
+
+
+
+# [0.2.0-alpha.25](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.24...v0.2.0-alpha.25) (2020-02-10)
+
+**Note:** Version bump only for package @rocket.chat/fuselage
+
+
+
+
+
+# [0.2.0-alpha.24](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.23...v0.2.0-alpha.24) (2020-02-09)
+
+**Note:** Version bump only for package @rocket.chat/fuselage
+
+
+
+
+
+# [0.2.0-alpha.23](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.22...v0.2.0-alpha.23) (2020-02-07)
+
+**Note:** Version bump only for package @rocket.chat/fuselage
+
+
+
+
+
+# [0.2.0-alpha.22](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.21...v0.2.0-alpha.22) (2020-02-07)
+
+
+### Bug Fixes
+
+* Select/MultiSelect focus submitting form ([#134](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/134)) ([630e622](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/630e622d8535fcaa06a8a736c6e0d273b450b739))
+* UIKit stopPropagation ([#137](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/137)) ([86939ea](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/86939eaae182d6554879468745780d83bc977b9e))
+
+
+
+
+
+# [0.2.0-alpha.21](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.20...v0.2.0-alpha.21) (2020-02-05)
+
+**Note:** Version bump only for package @rocket.chat/fuselage
+
+
+
+
+
+# [0.2.0-alpha.20](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.19...v0.2.0-alpha.20) (2020-02-01)
+
+
+### Bug Fixes
+
+* Modal background ([#126](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/126)) ([10b7202](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/10b72027cbb72d595d9aab8ffd37471eb7b8dac1))
+* Modal's margins and AnimatedVisibility initial state ([#128](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/128)) ([4357f82](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/4357f82250cad21a72b9a3332c0a31f4874e302a))
+* Remove visibilty on transition ([#129](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/129)) ([90c5301](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/90c5301a0067d00c73a249ca2c9d45206cb7ac05))
+
+
+
+
+
+# [0.2.0-alpha.19](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.18...v0.2.0-alpha.19) (2020-01-13)
+
+
+### Bug Fixes
+
+* InputBox sizing ([#124](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/124)) ([eb82e74](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/eb82e74784bf67655e07be4443672ed2e5a38728))
+* Position of inputs on selection buttons ([#123](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/123)) ([e446c89](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/e446c89d3a0c80e01266c373242d1197ba2950dc))
+
+
+### Features
+
+* Position, Modal, Options components ([#116](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/116)) ([af1916a](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/af1916a22c677939adda04fe417dafe406292762)), closes [#117](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/117)
+
+
+
+
+
 # [0.2.0-alpha.18](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.2.0-alpha.17...v0.2.0-alpha.18) (2019-12-31)